PresentChatsコンポーネントを新規作成し、index.tsxで利用するように修正
components/PresentChats.tsx
- useState, useRefをインポート
- useChatGPTをインポート
- PresentChatsコンポーネントを定義
- ボタンのクリックイベントを処理するhandleButtonClick関数を定義
- テキストエリアのフォーカスアウトイベントを処理するhandleBlur関数を定義
- ボタンとテキストエリアを表示するJSXを返す

pages/index.tsx
- PresentChatsコンポーネントをインポート
- PresentChatsコンポーネントを表示するように修正
